+++ tests/unit/test_discretisations/test_discretisation.py
@@ -144,9 +144,10 @@ class TestDiscretise(unittest.TestCase):
 
         # bounds with an InputParameter
         a = pybamm.InputParameter("a")
-        v = pybamm.Variable("v", domain=whole_cell, bounds=(0, a))
+        b = pybamm.InputParameter("b")
+        v = pybamm.Variable("v", domain=whole_cell, bounds=(a, b))
         disc.set_variable_slices([v])
-        np.testing.assert_array_equal(disc.bounds[0], [0] * 100)
+        np.testing.assert_array_equal(disc.bounds[0], [-np.inf] * 100)
         np.testing.assert_array_equal(disc.bounds[1], [np.inf] * 100)
 
     def test_process_symbol_base(self):